Refactoring Guice tests server/web
