+++ project/plugins.sbt
@@ -41,7 +41,3 @@ addSbtPlugin("com.typesafe.sbt" % "sbt-mocha" % "1.0.0")
 addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.8.2")
 addSbtPlugin("net.ground5hark.sbt" % "sbt-css-compress" % "0.1.3")
 addSbtPlugin("net.ground5hark.sbt" % "sbt-closure" % "0.1.3")
-
-
-// util
-addSbtPlugin("au.com.onegeek" %% "sbt-dotenv" % "2.1.146")
