+++ server/src/main/scala/org/ada/server/services/GuicePlayServerTestApp.scala
@@ -1,0 +1,12 @@
+package org.ada.server.services
+
+import play.api.Application
+import play.api.inject.guice.GuiceApplicationBuilder
+
+object GuicePlayServerTestApp extends GuicePlayTestApp {
+
+  override def apply(moduleNames: Seq[String] = Nil, excludeModules: Seq[String] = Nil): Application = {
+    new GuiceApplicationBuilder()
+      .configure("play.modules.enabled" -> getModules(moduleNames, excludeModules)).build
+  }
+}
