+++ server/src/test/scala/stats/MetricMDSTest.scala
@@ -2,7 +2,7 @@ package stats
 
 import akka.stream.scaladsl.Source
 import org.scalatest._
-import org.ada.server.services.{StatsService, GuicePlayTestApp}
+import org.ada.server.services.{StatsService, GuicePlayServerTestApp}
 
 import scala.concurrent.Future
 import scala.util.Random
@@ -31,7 +31,7 @@ class MetricMDSTest extends AsyncFlatSpec with Matchers with ExtraMatchers {
   private val randomInputSize = 1000
   private val precision = 0.001
 
-  private val injector = GuicePlayTestApp(addPlayCache = false).injector
+  private val injector = GuicePlayServerTestApp().injector
   private val statsService = injector.instanceOf[StatsService]
 
   "Metric MDS" should "match the static example" in {
